make sure we can generate NDJSON like list of resources to create/add to Fasten source when creating manually. Adding ReferenceUri identifier for referencing across sources.
