+++ frontend/src/app/pipes/reference-uri.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { ReferenceUriPipe } from './reference-uri.pipe';
+
+describe('ReferenceUriPipe', () => {
+  it('create an instance', () => {
+    const pipe = new ReferenceUriPipe();
+    expect(pipe).toBeTruthy();
+  });
+});
